refac: 위험운전 행동 분석 시스템 파일 충돌 해결 및 수정
- 위험운전 행동 집계 API 구현 (totalCounts)
  * DrivingAccumulatedStats 기반 실시간 집계
  * 급제동, 급가속, 차선변경 총합 제공
  * MilestoneItem 연결을 통한 리포트별 조회

- 시간대별 패턴 분석 시스템 구현 (drivingPattern)
  * 요일별(월~일) × 시간대별(5구간) 패턴 분석
  * Dominant Point 알고리즘으로 최다 발생 시간대 추출
  * 개인화된 운전 패턴 코멘트 생성
  * Mock 데이터 기반 (향후 Athena 쿼리 연동 예정)

- 이전 리포트 대비 증감 분석 구현 (compare)
  * 이전 리포트 자동 조회 (reportId - 1)
  * 정확한 증감률 계산 (소수점 2자리)
  * Edge case 처리 (이전 데이터 없음, 0값 등)

- 종합 분석 API 엔드포인트 제공
  * GET /api/driving-analysis/reports/{reportId}/behavior
  * 3개 태스크 통합 응답 구조
  * 에러 처리 및 기본값 제공

구현된 컴포넌트:
- BehaviorReportController, BehaviorReportService
- BehaviorPatternAnalyzer, BehaviorCompareAnalyzer
- BehaviorTotalCountsRepository, BehaviorPatternRepository
- 관련 DTO 및 Projection 인터페이스
- 단위 테스트 커버리지 포함

+++ src/main/java/com/smooth/driving_analysis_service/reports/accident_reaction/service/AccidentReactionWindowAnalyzerImpl.java
@@ -1,6 +1,7 @@
 package com.smooth.driving_analysis_service.reports.accident_reaction.service;
 
 import com.smooth.driving_analysis_service.driving.service.AthenaQueryService;
+import com.smooth.driving_analysis_service.reports.accident_reaction.dto.response.Reaction;
 import com.smooth.driving_analysis_service.reports.accident_reaction.support.DrivingEventType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -65,7 +66,7 @@ public class AccidentReactionWindowAnalyzerImpl implements AccidentReactionWindo
             log.debug("Reaction found: eventType={}, reactionMs={}, decelOrStop={}, evasiveManeuver={}", 
                     eventType, reactionMs, decelOrStop, evasiveManeuver);
             
-            return new Reaction(true, reactionMs, eventType, decelOrStop, evasiveManeuver);
+            return new Reaction(true, (long) reactionMs, eventType, decelOrStop, evasiveManeuver);
             
         } catch (Exception e) {
             log.error("Failed to analyze reaction window: userId={}, drivingId={}", userId, drivingId, e);
